% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_filter.R
\name{baseline_filter}
\alias{baseline_filter}
\title{Baseline filter}
\usage{
baseline_filter(shp, position, out_points)
}
\arguments{
\item{shp}{Shapefile. A point format shapefile with the intersections between each transect and the coastlines.}

\item{position}{character. The baseline position in relation to the coastlines. The options are: offshore, put "OFF"; onshore, put "ON"; or intermediate, put "MIX".}

\item{out_points}{Integer. Output name for the resulting shapefile.}
}
\value{
\itemize{
   \item{A shapefile with the filtered intersection points}
}
}
\description{
\code{baseline_filter} allows to filter the original interesections and select only the closest intersection between each transect and coastline.
}
\details{
The
}
\examples{
# Load data

}
