% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coast_rates.R
\name{coast_rates}
\alias{coast_rates}
\title{Coast Rates}
\usage{
coast_rates(inter_dist, normals, table, out_name)
}
\arguments{
\item{inter_dist}{Shapefile. A point format shapefile with the intersections between each transect and the coastlines.}

\item{normals}{Shapefile. A polyline shapefile with all normal lines included in the study site.}

\item{table}{CSV. A table with the information about the coastlines dates in format (dd/mm/yyyy) and the associated uncertainty to each coastline in meters. The column names should be "Date" and "Uncertainty".}

\item{out_name}{Integer. Output name for the resulting shapefile with the rates to each transect.}
}
\value{
\itemize{
   \item{A polyline shapefile with all rates associated to each transect.}
   \item{A table in png format with the central tendency stats for each parameter.}
}
}
\description{
\code{coast_rates} estimate the main parameters about the coastal variations.
}
\details{
The
}
\examples{
# Load data

}
