% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coast_var.R
\name{coast_var}
\alias{coast_var}
\title{Coastline Variations}
\usage{
coast_var(shp, normals, table, position, out_points, out_name)
}
\arguments{
\item{shp}{Shapefile. A point format shapefile with the intersections between each transect and the coastlines.}

\item{normals}{Shapefile. A polyline shapefile with all normal lines included in the study site.}

\item{table}{CSV. A table with the information about the coastlines dates in format (dd/mm/yyyy) and the associated uncertainty to each coastline in meters. The column names should be "Date" and "Uncertainty".}

\item{position}{character. The baseline position in relation to the coastlines. The options are: offshore, put "OFF"; onshore, put "ON"; or intermediate, put "MIX".}

\item{out_points}{Integer. Output name for the filter intersection points shapefile.}

\item{out_name}{Integer. Output name for the resulting shapefile with the rates to each transect.}
}
\value{
\itemize{
   \item{A shapefile with the filtered intersection points}
   \item{A polyline shapefile with all rates associated to each transect.}
   \item{A table in png format with the central tendency stats for each parameter.}
}
}
\description{
\code{coast_var} allows to filter the original interesections and select only the closest intersection between each transect and coastline.
}
\details{
The
}
\examples{
# Load data

}
