% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coast_rates.R
\name{coast_rates}
\alias{coast_rates}
\title{Calculate the coastal variation rates}
\usage{
coast_rates(inter_dist, normals, table, out_name)
}
\arguments{
\item{inter_dist}{Shapefile. A point format shapefile with the intersections between each transect and the shoreline.}

\item{normals}{Shapefile. A polyline format shapefile with all normal lines (transects) included in the study site.}

\item{table}{CSV. A table with the information about the coastlines dates in format (dd/mm/yyyy) and the associated uncertainty to each coastline in meters. The column names should be "Date" and "Uncertainty".}

\item{out_name}{Integer. Output name for the resulting shapefile with the rates of each transect.}
}
\value{
\itemize{
   \item{A polyline shapefile with all rates associated to each transect.}
   \item{A table in png format with the central tendency stats for each parameter.}
}
}
\description{
\code{coast_rates} estimate the main parameters about the coastal variations, as NSM, LRR or WLR.
}
\details{
This function is part of ODSAS method (Gomez-Pazo et al., \emph{2022}) and estimates the main five key measures for each individual transect and also produce aggregates
for all transects identically as within the DSAS tool (Himmelstoss et al., 2018).
The rates and their units are explained in detail in \code{CoastCR} documentation.
}
\examples{
library(sf)
library(CoastCR)

#Normal lines shapefile
normals <- st_read(system.file("./extdata/normals.shp", package = "CoastCR"))

#Table with dates and associated uncertainty
table <- read.csv(system.file("./extdata/table_coastlines.csv", package = "CoastCR"))

#Filtered intersections shapefile
inter_dist <- st_read(system.file("./extdata/dist.shp", package = "CoastCR"))

#Define output name
out_name <- "./normals_coast_rates.shp"

coast_rates(inter_dist, normals, table, out_name)


}
\references{
{
Gomez-Pazo, A., Payo, A., Paz-Delgado, M.V., Delgadillo-Calzadilla, M.A. (2022)
\emph{Open Digital Shoreline Analysis System: ODSAS v1.0}
Journal of Marine Science and Engineering, 10, 26.

Himmelstoss, E.A., Henderson, R.E., Kratzmann, M.G., Farris, A.S. (2018)
\emph{Digital Shoreline Analysis System (DSAS) version 5.0 user guide}
US Geological Survey Open File Report 2018-1179, 110 pp
}
}
\seealso{
\code{\link{baseline_filter}}; \code{\link{coast_var}}
}
